Create calculator
for formatting arguments ,first i write the code in string.xml . Then i write the above the code in activity_main.xml for portrait mode in mobile for better satisfaction. To give the button pressing effect i have created 3 xml file . By this the design of scientific calculator is finished and finally we have to give our java code for its functioning.